+++ packages/hub/src/kernel/meta/user-envelope/storage.ts
@@ -16,10 +16,10 @@
  *
  * @module
  */
-import type { NoydbStore, EncryptedEnvelope } from '../../kernel/types.js'
-import { NOYDB_FORMAT_VERSION } from '../../kernel/types.js'
-import { encrypt, decrypt } from '../../kernel/enclave/crypto.js'
-import { ConflictError } from '../../kernel/errors.js'
+import type { NoydbStore, EncryptedEnvelope } from '../../types.js'
+import { NOYDB_FORMAT_VERSION } from '../../types.js'
+import { encrypt, decrypt } from '../../enclave/crypto.js'
+import { ConflictError } from '../../errors.js'
 import {
   USER_ENVELOPE_COLLECTION,
   USER_ENVELOPE_MAX_BYTES,
