+++ packages/hub/src/kernel/meta/user-envelope/types.ts
@@ -5,11 +5,11 @@
  *
  * @module
  */
-import { NoydbError } from '../../kernel/errors.js'
+import { NoydbError } from '../../errors.js'
 
 /**
  * Thin reader view of a user envelope. The on-disk shape is the standard
- * {@link import('../../kernel/types.js').EncryptedEnvelope}; this is what callers
+ * {@link import('../../types.js').EncryptedEnvelope}; this is what callers
  * see after the storage layer has decrypted the payload.
  *
  * Hub commits to the `keyringId` ⇔ `userId` identity and the `_v` / `_ts`
